Move body and boundary to parametric module
